fix: don't lockup in React strict mode when using DragDropManager
Avoid passing the manager to the state initializer, as this causes issues with React strict mode. Instead, rely on the useEffect callback, already defined.

The \`manager\` arg passed back to the initializer function will now always be undefined, making it redundant.

It's still possible to cause strict mode to fail when using your own initalizer function to define a manager outside of dnd-kit, and pass it into the DragDropProvider.